perf(image): sfc to tsx
